shiro dto       test 改为普通项目

+++ spring-example-shiro/src/main/java/com/wyb/shiro/service/MenuService.java
@@ -9,5 +9,5 @@ import java.util.List;
  */
 public interface MenuService {
 
-    List<MenuDo> getByUserId(Integer id);
+    List<MenuDo> getByUserId(Long id);
 }
